@@ -1,12 +1,12 @@
 // ----------- import Packs
-import React from 'react';
-import { Provider } from 'react-redux';
-import { PersistGate } from 'redux-persist/integration/react';
+import React from "react";
+import { Provider } from "react-redux";
+import { PersistGate } from "redux-persist/integration/react";
 
 // ----------- import Internals
-import { store, persistor } from './config/redux/reduxConfig';
+import { store, persistor } from "./config/redux/reduxConfig";
 
-import AllApp from './project/AllApp';
+import AllApp from "./project/screens/A0_AllApp";
 
 // ----------- set Default Component
 export default () => (
@@ -16,3 +16,8 @@ export default () => (
     </PersistGate>
   </Provider>
 );
+
+// export default () => {
+//   console.log("aqui!!!");
+//   return <div>oi 3</div>;
+// };
