@@ -1,8 +1,8 @@
 // import Packages
 import firebase from './packs.web';
 
-import { fbWebConfig } from '../../appInfo';
-
+import appInfo from '../../appInfo';
+const { fbWebConfig } = appInfo;
 // const envConfig = process.env.NODE_ENV === 'production' ? prodConfig : devConfig;
 
 export const fbConfig =
