@@ -1,5 +1,5 @@
 export default {
-  A0: 'A0_AllAp',
+  A0: 'A0_AllApp',
   A0b: 'A0b_TempUsers',
   A1: 'A1_SignIn',
   A2: 'A2_Terms',
